+++ Reference Implementation/TradingApp2/DataModel/DataModels/TransactionViewModel.cs
@@ -80,7 +80,7 @@ namespace TradingApp2.TradeLibrary.DataModels
             }
         }
 
-        public int Id { get { return _model.id; } }
+        public long Id { get { return _model.id; } }
         public int AccountId { get { return _model.accountId; } }
         public string Type { get { return _model.type; } }
         public string Side { get { return _model.side; } }
